Moved generate/execute modules to tkltest.generate/execute.unit package; related code and test updates (#151)
