+++ tkltest/execute/unit/execute.py
@@ -19,10 +19,10 @@ import shutil
 
 import toml
 
-from tkltest.execute import compare_coverage
+from tkltest.execute.unit import compare_coverage
 from tkltest.util import constants, build_util, command_util, dir_util, config_util, coverage_util
 from tkltest.util.logging_util import tkltest_status
-from tkltest.execute.coverage_html_writer import CoverageStatisticsHtmlWriter
+from tkltest.execute.unit.coverage_html_writer import CoverageStatisticsHtmlWriter
 
 
 def process_execute_command(args, config):
